requirements.txt
# Core Dependencies for Biosensor Data Capture Tool
# Cross-platform base requirements

pyautogui==0.9.54
PyWinCtl==0.0.50
Pillow==10.4.0
requests==2.32.3
opencv-python==4.10.0.84
numpy==2.1.1
mss==9.0.1
python-dotenv==1.0.0
psutil==5.9.8

# UI Libraries
PyQt5==5.15.10
qtawesome==1.3.1
qdarkstyle==3.2.3
PyQt5-stubs==5.15.6.0

# For platform-specific dependencies:
# Windows: pip install -r windows_requirements.txt
# Linux: pip install -r linux_requirements.txt